requirements.txt
flask
flask-cors
yfinance
python-dotenv
apscheduler
websockets
polygon-api-client
redis
rq
PyJWT
flask-socketio